requirement.txt
google-generativeai
streamlit
python-dotenv

-e .